Renames per PR review suggestions
