+++ src/numbers/int128.js
@@ -1,6 +1,12 @@
 import { LargeInt } from 'js-xdr';
 
 export class Int128 extends LargeInt {
+  /**
+   * Construct a signed 128-bit integer that can be XDR-encoded.
+   *
+   * @param  {Array<number|bigint|string>}  args - one or more slices to encode
+   *     in big-endian format (i.e. earlier elements are higher bits)
+   */
   constructor(...args) {
     super(args);
   }
