+++ src/numbers/uint256.js
@@ -1,6 +1,12 @@
 import { LargeInt } from 'js-xdr';
 
 export class Uint256 extends LargeInt {
+  /**
+   * Construct an unsigned 256-bit integer that can be XDR-encoded.
+   *
+   * @param  {Array<number|bigint|string>}  args - one or more slices to encode
+   *     in big-endian format (i.e. earlier elements are higher bits)
+   */
   constructor(...args) {
     super(args);
   }
